# Use separate connection cache for tpu forward, and use connection pool size = 1