@behzadnouri 2024-04-19
> There is not expected to have that many forwarded packets. This will help reduce the load on the leader.

right, but often times the experiment results do not match expectations.
So might be good to do some a/b testing this code on some to confirm.

Another thing, because of these `static`s:
https://github.com/anza-xyz/agave/blob/0556020c6/quic-client/src/quic_client.rs#L68-L76
we are practically sharing resources between these two. is that correct?