@lijunwangs 2024-04-19
I do not expect material impacts on tpu performance using bench-tps tool, as forward is a small part of the overall performance.
WIth the change:

[2024-04-19T18:48:26.108934546Z INFO  solana_bench_tps::bench] http://35.233.177.221:8899 |      63840.37 | 291780 
[2024-04-19T18:48:26.108940767Z INFO  solana_bench_tps::bench] 
    Average max TPS: 63840.37, 0 nodes had 0 TPS
[2024-04-19T18:48:26.108943890Z INFO  solana_bench_tps::bench] 
    Highest TPS: 63840.37 sampling period 1s max transactions: 291780 clients: 1 drop rate: 0.84
[2024-04-19T18:48:26.108947599Z INFO  solana_bench_tps::bench] 	Average TPS: 9288.707



[2024-04-19T18:58:12.785494563Z INFO  solana_bench_tps::bench] http://35.233.177.221:8899 |      41793.30 | 268613 
[2024-04-19T18:58:12.785498584Z INFO  solana_bench_tps::bench] 
    Average max TPS: 41793.30, 0 nodes had 0 TPS
[2024-04-19T18:58:12.785509050Z INFO  solana_bench_tps::bench] 
    Highest TPS: 41793.30 sampling period 1s max transactions: 268613 clients: 1 drop rate: 0.87
[2024-04-19T18:58:12.785512202Z INFO  solana_bench_tps::bench] 	Average TPS: 8148.3394



Without it


[2024-04-19T19:08:36.258726268Z INFO  solana_bench_tps::bench] http://35.233.177.221:8899 |      52674.86 | 347864 
[2024-04-19T19:08:36.258731874Z INFO  solana_bench_tps::bench] 
    Average max TPS: 52674.86, 0 nodes had 0 TPS
[2024-04-19T19:08:36.258735780Z INFO  solana_bench_tps::bench] 
    Highest TPS: 52674.86 sampling period 1s max transactions: 347864 clients: 1 drop rate: 0.80
[2024-04-19T19:08:36.258740146Z INFO  solana_bench_tps::bench] 	Average TPS: 10952.888


[2024-04-19T19:19:05.074011907Z INFO  solana_bench_tps::bench] Token balance: 29703545150
[2024-04-19T19:19:05.074045136Z INFO  solana_bench_tps::bench]  Node address        |       Max TPS | Total Transactions
[2024-04-19T19:19:05.074047078Z INFO  solana_bench_tps::bench] ---------------------+---------------+--------------------
[2024-04-19T19:19:05.074049651Z INFO  solana_bench_tps::bench] http://35.233.177.221:8899 |      51925.07 | 284671 
[2024-04-19T19:19:05.074055000Z INFO  solana_bench_tps::bench] 
    Average max TPS: 51925.07, 0 nodes had 0 TPS
[2024-04-19T19:19:05.074058135Z INFO  solana_bench_tps::bench] 
    Highest TPS: 51925.07 sampling period 1s max transactions: 284671 clients: 1 drop rate: 0.78
[2024-04-19T19:19:05.074061081Z INFO  solana_bench_tps::bench] 	Average TPS: 8408.443