@lijunwangs 2024-04-19
I also use --tpu-connection-pool-size parameter when running bench-tps to test the difference between 1 and 4. Actually I am not seeing a lot of difference. Our QOS stuff change, Quinn change and tokio change probably make the connection count less critical in the performance now as compared like 2 years ago.