@lijunwangs 2024-04-20
> Have we tested the practical impact of this? Wouldn't this just increase the global connection limit to any given remote node (the total number of connections a node might open to another node) by 1?

replied below with performance numbers