@lijunwangs 2024-04-21
> This code is creating two caches but:
> 
> * one is not used by staked nodes, and the other is not used by unstaked or rpc nodes. so pretty ugly if not wasteful in that regard.
> * the caches are only different in pool size (also one is warmed and the other is not).
> * there is some complexity in how the two caches still interfere with each other through these `static`s:
>   https://github.com/anza-xyz/agave/blob/0556020c6/quic-client/src/quic_client.rs#L68-L76
> 
> I think a much better solution is that:
> 
> * keep only one connection cache just like before.
> * adjust the default value of `tpu-connection-pool-size` so that if the node is not rpc, it defaults to 1:
>   https://github.com/anza-xyz/agave/blob/ff7953120/validator/src/cli.rs#L877-L882

1. I think you are assuming staked nodes are RPC nodes are mutually exclusive. Is that true? What is the evidence? Our code does not prevent a staked node to be a RPC node.
2. Not sure what is the the wasteful. The entries in the cache will not be increased. Note the key is SocketAddr which are different for tpu and tpu_forward. I do not think it has additional memory usage. 
3. I am not solving the interference tpu and tpu_forward you pointed out here. of https://github.com/anza-xyz/agave/blob/0556020c6/quic-client/src/quic_client.rs#L68-L76. This change does not make any more better or worse in this regard.