@behzadnouri 2024-04-21
> I think you are assuming staked nodes are RPC nodes are mutually exclusive. Is that true? What is the evidence? Our code does not prevent a staked node to be a RPC node.

it is strongly discouraged for staked node to expose rpc interface due to security issues.